Update django documentation and include image in the example
