+++ demo/djangoproject/utils.py
@@ -1,0 +1,25 @@
+# encoding: utf-8
+
+from __future__ import unicode_literals
+from django.template import Context, Template
+
+
+def extract_request_variables(request):
+
+    page_size = request.POST.get('page_size', 'letter')
+    page_orientation = request.POST.get('page_orientation', 'portrait')
+
+    pagesize = "%s %s" % (
+        page_size, page_orientation
+    )
+
+    template = Template(request.POST.get('data', ''))
+    data = template.render(Context({}))
+    return {
+        'pagesize': pagesize,
+        'data': data,
+        'page_orientation': page_orientation,
+        'page_size': page_size,
+        'example_number': request.POST.get("example_number", '1'),
+        'border': request.POST.get('border', '')
+    }
