+++ demo/djangoproject/views.py
@@ -1,37 +1,51 @@
 # coding: utf-8
 
-import datetime
+import os
 
+from django.conf import settings
 from django.http import HttpResponse
 from django.shortcuts import render
-from django.template import Context, Template
-from django.template.loader import get_template, render_to_string
+from django.template import Context
+from django.template.loader import get_template
 from xhtml2pdf import pisa
+
+
+try:  # python2 and python3
+    from .utils import extract_request_variables
+except:
+    from utils import extract_request_variables
 
 
 def index(request):
     return render(request, 'index.html')
 
 
-def extract_request_variables(request):
+def link_callback(uri, rel):
+    """
+    Convert HTML URIs to absolute system paths so xhtml2pdf can access those
+    resources
+    """
+    # use short variable names
+    sUrl = settings.STATIC_URL      # Typically /static/
+    sRoot = settings.STATIC_ROOT    # Typically /home/userX/project_static/
+    mUrl = settings.MEDIA_URL       # Typically /static/media/
+    # Typically /home/userX/project_static/media/
+    mRoot = settings.MEDIA_ROOT
 
-    page_size = request.POST.get('page_size', 'letter')
-    page_orientation = request.POST.get('page_orientation', 'portrait')
+    # convert URIs to absolute system paths
+    if uri.startswith(mUrl):
+        path = os.path.join(mRoot, uri.replace(mUrl, ""))
+    elif uri.startswith(sUrl):
+        path = os.path.join(sRoot, uri.replace(sUrl, ""))
+    else:
+        return uri  # handle absolute uri (ie: http://some.tld/foo.png)
 
-    pagesize = "%s %s" % (
-        page_size, page_orientation
-    )
-
-    template = Template(request.POST.get('data', ''))
-    data = template.render(Context({}))
-    return {
-        'pagesize': pagesize,
-        'data': data,
-        'page_orientation': page_orientation,
-        'page_size': page_size,
-        'example_number': request.POST.get("example_number", '1'),
-        'border': request.POST.get('border', '')
-    }
+    # make sure that file exists
+    if not os.path.isfile(path):
+        raise Exception(
+            'media URI must start with %s or %s' % (sUrl, mUrl)
+        )
+    return path
 
 
 def render_pdf(request):
@@ -49,7 +63,8 @@ def render_pdf(request):
         response['Content-Disposition'] = 'attachment; filename="report.txt"'
         response.write(html)
     else:
-        pisaStatus = pisa.CreatePDF(html, dest=response)
+        pisaStatus = pisa.CreatePDF(
+            html, dest=response, link_callback=link_callback)
         if pisaStatus.err:
             return HttpResponse('We had some errors <pre>' + html + '</pre>')
     return response
